refactor: useJoinMutation 내부에서 popup 로직 처리
