+++ src/components/templates/login/JoinForm/index.tsx
@@ -1,21 +1,23 @@
 import { useRouter } from 'next/router';
+import { useCallback, useEffect, useState } from 'react';
 import styled from 'styled-components';
-import { useForm, FormProvider } from 'react-hook-form';
+import { useForm, FormProvider, SubmitHandler } from 'react-hook-form';
 import DefaultForm from '@src/components/common/DefaultForm';
-import { JoinFormData, StepType } from '@src/types/joinTypes';
 import EmailForm from '@src/components/organism/join/EmailForm';
 import NameForm from '@src/components/organism/join/NameForm';
-import { useEffect, useState } from 'react';
+import { JoinFormData, StepType } from '@src/types/joinTypes';
 import IdForm from '@src/components/organism/join/IdForm';
 import PasswordForm from '@src/components/organism/join/PasswrodForm';
 import AgreeForm from '@src/components/organism/join/AgreeForm';
+import useJoinMutation from '@src/hooks/apis/useJoinMutation';
 
 function JoinForm() {
+  const { mutate } = useJoinMutation();
   const [step, setStep] = useState<StepType>('agree');
   const methods = useForm<JoinFormData>();
   const { query } = useRouter();
 
-  useEffect(() => {
+  const controllFunnel = useCallback(() => {
     const funnelStep = query['funnel-step'];
 
     if (!funnelStep || funnelStep.includes('agree')) {
@@ -33,7 +35,12 @@ function JoinForm() {
     }
   }, [query]);
 
-  const onValid = () => {};
+  useEffect(controllFunnel, [controllFunnel]);
+
+  const onValid: SubmitHandler<JoinFormData> = (data) => {
+    const { id, pw, email, name, nickname } = data;
+    mutate({ id, pw, email, name, nickname });
+  };
   const onInvalid = () => {};
 
   return (
