+++ src/hooks/apis/useJoinMutation/index.ts
@@ -4,9 +4,13 @@ import { PATH } from '@src/constants/urls';
 import { useMutation } from '@tanstack/react-query';
 import axios, { AxiosResponse } from 'axios';
 import { useRouter } from 'next/router';
+import { POPUP_MESSAGE } from '@src/constants/message';
+import { POPUP_DURATION } from '@src/constants/duration';
+import usePopup from '@src/hooks/usePopup';
 
 function useJoinMutation() {
   const { push } = useRouter();
+  const { openAndDeletePopup } = usePopup();
 
   const mutationFn = (params: JoinReqData) => axios.post(JOIN_API, params);
   const onSuccess = (data: AxiosResponse<JoinResData>) => {
@@ -15,6 +19,10 @@ function useJoinMutation() {
       return;
     }
 
+    openAndDeletePopup({
+      message: POPUP_MESSAGE.joinSuccess,
+      duration: POPUP_DURATION.medium,
+    });
     push(PATH.user.login.url);
   };
